Small clarification to edition compatibility
The old text ("The Rust compiler can link crates
of any editions together") made it sound like old
compilers would somehow magically be able to support
language features from newer editions.

The goal of this new text is to make it more clear
that a given compiler release will only be able to
work with the specific editions that it supports.