fixes #277: mention rust 1.0 release month and year
In the "What are Editions?" section, note that Rust 1.0 was released
in May 2015, since it answers a likely question in the mind of the
reader considering the preceding commentary referencing that time
frame. Writing in Q3 2022, noting the date also helps establish the
Rust project's credible commitment to the rule of supporting stable
features in perpetuity. It is not something that was just thought up
recently -- the project has been adhering to it for more than seven
years.

Signed-off-by: Alan D. Salewski <ads@salewski.email>