rust-2018/path-changes: updates for Rust 1.72
Since rust-lang/rust#112086, paths in `use` statements have the same shadowing
rules as other paths, at least for the simple cases we're talking about in
this guide.

The case of an external crate having the same name as a local module still
seems worth mentioning, so I've put it in the "Extern crate paths" section.