Align page title with summary title
Aside from aligning it with the summary entry, another reason to say
"never type fallback change" is that if we were to say "to itself", we
should say "never type fall back to itself", grammatically.  But it
seems better here to just use the noun form, so we'll be consistent
with that.