Make an editing pass and make all tests work
On the page for the never type fallback change, we've made a general
copyediting pass.  As part of this, we've also fixed all of the tests
such that they are checked and not simply ignored.

Strangely, some of the longer explanations that had been added to the
ignored tests were causing weird formatting issues in the rendered
output.  The example blocks weren't taking up the full width, and text
was flowing around them.  Making the tests work instead, and removing
these explanations, resolves this issue.