Make T Sync in the SyncUnsafeCell example
I initially did not include it because I saw a few comments from people
concerned with std's decision because they had !Sync types that they
believed would be safe to use. However, I think it would be good to
include the bound here, just for the concern that someone might
copy-paste this example. If they really know what they are doing, they
can customize it.